Megacommit: Refactor installation process
- Replaced SSCM installation with BepInEx installation in ReloadConfig function.
- Added RestartBackend function to facilitate backend restarts.
- Enhanced and renamed varous packages
- persisted isFirstTimeSetup
- Updated web server startup to display first-time setup message if applicable.
